Flytt to valideringsfunksjoner inn i valideringsfil
