+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vedtak/vedtaksperiode/VedtakPeriodeValidering.kt
@@ -1,10 +1,41 @@
 package no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode
 
+import no.nav.familie.ba.sak.common.FunksjonellFeil
+import no.nav.familie.ba.sak.common.TIDENES_MORGEN
 import no.nav.familie.ba.sak.common.tilKortString
+import no.nav.familie.ba.sak.kjerne.beregning.SatsService
 import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.VedtakBegrunnelseType
+import no.nav.familie.ba.sak.kjerne.vedtak.domene.VedtaksperiodeMedBegrunnelser
 import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.domene.UtvidetVedtaksperiodeMedBegrunnelser
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import java.time.LocalDate
+
+fun validerSatsendring(
+    fom: LocalDate?,
+    harBarnMedSeksårsdagPåFom: Boolean,
+) {
+    val satsendring = SatsService.finnSatsendring(fom ?: TIDENES_MORGEN)
+
+    if (satsendring.isEmpty() && !harBarnMedSeksårsdagPåFom) {
+        throw FunksjonellFeil(
+            melding = "Begrunnelsen stemmer ikke med satsendring.",
+            frontendFeilmelding = "Begrunnelsen stemmer ikke med satsendring. Vennligst velg en annen begrunnelse.",
+        )
+    }
+}
+
+fun validerVedtaksperiodeMedBegrunnelser(vedtaksperiodeMedBegrunnelser: VedtaksperiodeMedBegrunnelser) {
+    if ((vedtaksperiodeMedBegrunnelser.type == Vedtaksperiodetype.OPPHØR || vedtaksperiodeMedBegrunnelser.type == Vedtaksperiodetype.AVSLAG) && vedtaksperiodeMedBegrunnelser.harFriteksterUtenStandardbegrunnelser()) {
+        val fritekstUtenStandardbegrunnelserFeilmelding =
+            "Fritekst kan kun brukes i kombinasjon med en eller flere begrunnelser. " + "Legg først til en ny begrunnelse eller fjern friteksten(e)."
+        throw FunksjonellFeil(
+            melding = fritekstUtenStandardbegrunnelserFeilmelding,
+            frontendFeilmelding = fritekstUtenStandardbegrunnelserFeilmelding,
+        )
+    }
+}
+
 
 // Håpet er at denne skal kaste feil på sikt, men enn så lenge blir det for strengt. Logger for å se behovet.
 fun List<UtvidetVedtaksperiodeMedBegrunnelser>.validerPerioderInneholderBegrunnelser(
