--- src/main/kotlin/no/nav/familie/ba/sak/kjerne/vedtak/vedtaksperiode/VedtaksperiodeUtil.kt
@@ -1,32 +1,0 @@
-package no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode
-
-import no.nav.familie.ba.sak.common.FunksjonellFeil
-import no.nav.familie.ba.sak.common.TIDENES_MORGEN
-import no.nav.familie.ba.sak.kjerne.beregning.SatsService
-import no.nav.familie.ba.sak.kjerne.vedtak.domene.VedtaksperiodeMedBegrunnelser
-import java.time.LocalDate
-
-fun validerSatsendring(
-    fom: LocalDate?,
-    harBarnMedSeksårsdagPåFom: Boolean,
-) {
-    val satsendring = SatsService.finnSatsendring(fom ?: TIDENES_MORGEN)
-
-    if (satsendring.isEmpty() && !harBarnMedSeksårsdagPåFom) {
-        throw FunksjonellFeil(
-            melding = "Begrunnelsen stemmer ikke med satsendring.",
-            frontendFeilmelding = "Begrunnelsen stemmer ikke med satsendring. Vennligst velg en annen begrunnelse.",
-        )
-    }
-}
-
-fun validerVedtaksperiodeMedBegrunnelser(vedtaksperiodeMedBegrunnelser: VedtaksperiodeMedBegrunnelser) {
-    if ((vedtaksperiodeMedBegrunnelser.type == Vedtaksperiodetype.OPPHØR || vedtaksperiodeMedBegrunnelser.type == Vedtaksperiodetype.AVSLAG) && vedtaksperiodeMedBegrunnelser.harFriteksterUtenStandardbegrunnelser()) {
-        val fritekstUtenStandardbegrunnelserFeilmelding =
-            "Fritekst kan kun brukes i kombinasjon med en eller flere begrunnelser. " + "Legg først til en ny begrunnelse eller fjern friteksten(e)."
-        throw FunksjonellFeil(
-            melding = fritekstUtenStandardbegrunnelserFeilmelding,
-            frontendFeilmelding = fritekstUtenStandardbegrunnelserFeilmelding,
-        )
-    }
-}
